# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Permionics BD Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questionnaires:
    get:
      operationId: listQuestionnaires
      tags: [questionnaires]
      summary: List all saved questionnaires
      responses:
        "200":
          description: List of questionnaires
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Questionnaire"
    post:
      operationId: createQuestionnaire
      tags: [questionnaires]
      summary: Save a completed questionnaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Questionnaire"
  /questionnaires/{id}:
    get:
      operationId: getQuestionnaire
      tags: [questionnaires]
      summary: Get a single questionnaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questionnaire detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Questionnaire"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteQuestionnaire
      tags: [questionnaires]
      summary: Delete a questionnaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /settings/change-password:
    post:
      operationId: changePassword
      tags: [settings]
      summary: Change the admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AuthStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
        message:
          type: ["string", "null"]
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 6
...
    AnswerMap:
      type: object
      additionalProperties:
        type: string
    Questionnaire:
      type: object
      required: [id, clientName, sector, questions, answers, createdAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        sector:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionItem"
        answers:
          $ref: "#/components/schemas/AnswerMap"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    QuestionItem:
      type: object
      required: [id, question, type]
      properties:
        id:
          type: string
        question:
          type: string
        type:
          type: string
        options:
          type: array
          items:
            type: string
    QuestionnaireInput:
      type: object
      required: [clientName, sector, questions, answers]
      properties:
        clientName:
          type: string
          minLength: 1
        sector:
          type: string
          minLength: 1
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionItem"
        answers:
          $ref: "#/components/schemas/AnswerMap"
        notes:
          type: ["string", "null"]
...