# lib/api-zod/src/generated/types/questionItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Permionics BD Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionItem {
  id: string;
  question: string;
  type: string;
  options?: string[];
}